// src/types/qwallet-error.enum.ts
export enum QwalletErrorEnum {
  // Sub-account management
  SUBACCOUNT_ALREADY_EXISTS = 'SUBACCOUNT_ALREADY_EXISTS',
  SUBACCOUNT_CREATE_FAILED = 'SUBACCOUNT_CREATE_FAILED',
  FETCH_SUBACCOUNTS_FAILED = 'FETCH_SUBACCOUNTS_FAILED',
  FETCH_SUBACCOUNT_DETAILS_FAILED = 'FETCH_SUBACCOUNT_DETAILS_FAILED',
  SUBACCOUNT_UPDATE_FAILED = 'SUBACCOUNT_UPDATE_FAILED',
  FETCH_PARENT_ACCOUNT_FAILED = 'FETCH_PARENT_ACCOUNT_FAILED',

  // Wallets
  CREATE_USER_WALLET_FAILED = 'CREATE_USER_WALLET_FAILED',
  GET_USER_WALLETS_FAILED = 'GET_USER_WALLETS_FAILED',
  GET_USER_WALLET_FAILED = 'GET_USER_WALLET_FAILED',
  GET_PAYMENT_ADDRESS_FAILED = 'GET_PAYMENT_ADDRESS_FAILED',
  FETCH_PAYMENT_ADDRESSES_FAILED = 'FETCH_PAYMENT_ADDRESSES_FAILED',
  PAYMENT_ADDRESS_BY_ID_FETCH_FAILED = 'PAYMENT_ADDRESS_BY_ID_FETCH_FAILED',
  CREATE_PAYMENT_ADDRESS_FAILED = 'CREATE_PAYMENT_ADDRESS_FAILED',
  REENQUEUE_WALLET_ADDRESS_FAILED = 'REENQUEUE_WALLET_ADDRESS_FAILED',

  // Address validation
  VALIDATE_ADDRESS_FAILED = 'VALIDATE_ADDRESS_FAILED',

  // Withdrawals
  FETCH_WITHDRAWALS_FAILED = 'FETCH_WITHDRAWALS_FAILED',
  CANCEL_WITHDRAWAL_FAILED = 'CANCEL_WITHDRAWAL_FAILED',
  GET_WITHDRAWAL_FAILED = 'GET_WITHDRAWAL_FAILED',
  CREATE_WITHDRAWAL_FAILED = 'CREATE_WITHDRAWAL_FAILED',
  FETCH_WITHDRAWAL_BY_REFERENCE_FAILED = 'FETCH_WITHDRAWAL_BY_REFERENCE_FAILED',

  // Generic fallback
  QWALLET_INTERNAL_ERROR = 'QWALLET_INTERNAL_ERROR',
}